Move PointSelectionEvent handling to common class
