+++ iris-visualizer/src/main/java/cfa/vo/iris/visualizer/plotter/PlotPointSelectionListener.java
@@ -18,30 +18,18 @@ package cfa.vo.iris.visualizer.plotter;
 import cfa.vo.iris.visualizer.metadata.MetadataBrowserMainView;
 import uk.ac.starlink.ttools.plot2.task.PlotDisplay;
 import uk.ac.starlink.ttools.plot2.task.PointSelectionEvent;
-import uk.ac.starlink.ttools.plot2.task.PointSelectionListener;
 
 /**
  * MouseListener which allows users to click on the PlotDisplay and add the selected
  * point to the selection in the metadata browser.
  *
  */
-public class PlotPointSelectionListener extends StilPlotterMouseListener
-    implements PointSelectionListener 
+public class PlotPointSelectionListener extends StilPlotterPointSelectionListener
 {
-    
     @Override
-    public void pointSelected(PointSelectionEvent evt) {
+    public void handleSelection(int starTableIndex, int irow, PointSelectionEvent evt) {
         MetadataBrowserMainView mbView = this.getPlotterView().getMetadataBrowserView();
-        
-        long[] rows = evt.getClosestRows();
-        for (int i=0; i*2<rows.length; i++) {
-            // It is possible that the Event will contain multiple selected points, but
-            // we will only use the first point selected. See http://tinyurl.com/gq9o7te
-            if (rows[2*i] >= 0) {
-                mbView.addRowToSelection(i, (int) rows[2*i]);
-                return;
-            }
-        }
+        mbView.addRowToSelection(starTableIndex, irow);
     }
     
     @Override
